[Fix] Fix the warning that cause debug compile failure
Summary:
Fix the comparison warning that cause debug compile failure

Test Plan:
make DEBUG=1

DiffCamp Revision: 154328
Reviewed By: mwilliams
CC: myang, mwilliams, hphp-diffs@lists
Revert Plan:
OK

+++ src/runtime/base/shared/immutable_map.cpp
@@ -34,7 +34,7 @@ void ImmutableMap::add(ThreadSharedVariant *key, ThreadSharedVariant *val) {
   // NOTE: no check on duplication because we assume the original array has no
   // duplication
   int pos = m_curPos++;
-  ASSERT(pos <= m_capacity_mask);
+  ASSERT(pos <= (int)m_capacity_mask);
   m_buckets[pos].key = key;
   m_buckets[pos].val = val;
   if (key->is(KindOfInt64)) {
